[minor] always generate a thin _app.js
